Restructure source files
