+++ include/Basis/CMakeLists.txt
@@ -1,14 +1,9 @@
 target_sources(gqcp
         PRIVATE
-            CartesianExponents.hpp
-            CartesianGTO.hpp
-            GTOBasisSet.hpp
-            GTOShell.hpp
-            RSpinorBasis.hpp
-            ScalarBasis.hpp
-            ShellSet.hpp
-            TransformationMatrix.hpp
             transform.hpp
+            TransformationMatrix.hpp
         )
 
 add_subdirectory(Integrals)
+add_subdirectory(ScalarBasis)
+add_subdirectory(SpinorBasis)
