+++ include/Basis/ScalarBasis/CMakeLists.txt
@@ -1,0 +1,9 @@
+target_sources(gqcp
+        PRIVATE
+            CartesianExponents.hpp
+            CartesianGTO.hpp
+            GTOBasisSet.hpp
+            GTOShell.hpp
+            ScalarBasis.hpp
+            ShellSet.hpp
+        )
