+++ include/Basis/SpinorBasis/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(gqcp
+        PRIVATE
+            GSpinorBasis.hpp
+            RSpinorBasis.hpp
+        )
