+++ include/QCMethod/FukuiDysonAnalysis.hpp
@@ -17,7 +17,8 @@
 //
 #pragma once
 
-#include "Basis/RSpinorBasis.hpp"
+
+#include "Basis/SpinorBasis/RSpinorBasis.hpp"
 #include "CISolver/CISolver.hpp"
 #include "FockSpace/ProductFockSpace.hpp"
 #include "HamiltonianBuilder/FCI.hpp"
