+++ include/QCMethod/MullikenConstrainedFCI.hpp
@@ -18,7 +18,7 @@
 #pragma once
 
 
-#include "Basis/RSpinorBasis.hpp"
+#include "Basis/SpinorBasis/RSpinorBasis.hpp"
 #include "CISolver/CISolver.hpp"
 #include "FockSpace/ProductFockSpace.hpp"
 #include "HamiltonianBuilder/FCI.hpp"
