+++ src/Basis/Integrals/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(gqcp
+        PRIVATE
+            IntegralEngine.cpp
+        )
+
+add_subdirectory(Interfaces)
