+++ src/Basis/Integrals/Interfaces/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(gqcp
+        PRIVATE
+            LibcintInterfacer.cpp
+            LibintInterfacer.cpp
+        )
+
