+++ src/Basis/ScalarBasis/CMakeLists.txt
@@ -1,0 +1,7 @@
+target_sources(gqcp
+        PRIVATE
+            CartesianExponents.cpp
+            CartesianGTO.cpp
+            GTOBasisSet.cpp
+            GTOShell.cpp
+        )
