+++ src/Basis/ScalarBasis/GTOBasisSet.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
 // 
-#include "Basis/GTOBasisSet.hpp"
+#include "Basis/ScalarBasis/GTOBasisSet.hpp"
 
 #include "Basis/Integrals/Interfaces/LibintInterfacer.hpp"
