+++ tests/Basis/CMakeLists.txt
@@ -1,12 +1,9 @@
 list(APPEND test_target_sources
-        ${CMAKE_CURRENT_SOURCE_DIR}/CartesianExponents_test.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/CartesianGTO_test.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/GTOShell_test.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/LibintInterfacer_test.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/ScalarBasis_test.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/ShellSet_test.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/RSpinorBasis_test.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/TransformationMatrix_test.cpp
         )
+
+add_subdirectory(Integrals)
+add_subdirectory(ScalarBasis)
+add_subdirectory(SpinorBasis)
 
 set(test_target_sources ${test_target_sources} PARENT_SCOPE)
