+++ tests/Basis/Integrals/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory(Interfaces)
+
+set(test_target_sources ${test_target_sources} PARENT_SCOPE)
