+++ tests/Basis/ScalarBasis/CartesianGTO_test.cpp
@@ -19,7 +19,7 @@
 
 #include <boost/test/unit_test.hpp>
 
-#include "Basis/CartesianGTO.hpp"
+#include "Basis/ScalarBasis/CartesianGTO.hpp"
 
 
 BOOST_AUTO_TEST_CASE ( constructor ) {
