+++ tests/Basis/SpinorBasis/CMakeLists.txt
@@ -1,0 +1,5 @@
+list(APPEND test_target_sources
+        ${CMAKE_CURRENT_SOURCE_DIR}/RSpinorBasis_test.cpp
+        )
+
+set(test_target_sources ${test_target_sources} PARENT_SCOPE)
